Manual IP connection
This patch adds a new menu to add and remove IP addresses and hostnames,
and makes LanBackend read that list of addresses and send identity packets
to them. This bypasses the autodiscovery which usually happens via
broadcast packets, and allows the pairing and communication of devices even
if broadcast is not possible between them.

BUG: 326509
REVIEW: 121104