Show notifications that are present when pairing
Summary:
When the plugin is created all current notifications are sent to the other device. When the NotificationListener is not ready at that time this is delayed until it is. This was intended before, but did not work because the
requestAnswer flag was set on the network packages. The desktop ignored those packages. I see no purpose for the requestAnswer flag so I removed it.

Depends on D9622

Test Plan: Create some notifications. Restart the desktop or the app and look for the notifications on the desktop

Reviewers: #kde_connect, mtijink, albertvaka

Reviewed By: #kde_connect, mtijink, albertvaka

Subscribers: mtijink, #kde_connect

Tags: #kde_connect

Differential Revision: https://phabricator.kde.org/D9558