Lessen indent before "Send files" in main activity
Summary:
The "Send files" button in main acitivity if connected has a larger
indent then the other buttons. This happens on most pixel densities.
Reason was the larger icon the "Send files" button has.

With this patch, the icons all have the same pixel dimensions. It
solves the too large indent.

Screenshot (Because of translation, "Send files" is translated to
"Dateien senden"):

{F6136062}

Reviewers: #kde_connect, apol, nicolasfella

Reviewed By: #kde_connect, nicolasfella

Subscribers: nicolasfella, apol, kdeconnect

Tags: #kde_connect

Differential Revision: https://phabricator.kde.org/D14344