Add a Text Input plugin
This plugin allows the user to use more complex input methods, such as
the Google's Handwriting Keyboard for Japanese. Text is composed in a
text field and submitted by a button. Tested with English and Japanese.

Emoji don't work, but they don't work with the MousePad plugin either,
and I reused its event for sending text, so that bug is somewhere
deeper (maybe in the desktop software). Emoji are always sent as ??.