[SMS Plugin] Fix handling for addresses list of multi-target group messages
## Summary

Fix a logic error in the SMS plugin which caused it to skip the addresses of other targets when reporting group messages.

Notionally, this bug originated with the logic change in !197, where everything switched to using the android-smsmms library.

BUG: 464555

## Test Plan

### Before:
Group messages would be returned as single-target messages between you and the sender, meaning the desktop app would have a bad time showing them.

### After:
Group messages are returned with the appropriate list of addresses, the desktop app groups them appropriately.