SystemVolumePlugin: Do not call requestSinkList once per second
When playing music on Windows, we receive MPRIS player updates ~once per
second. This calls the MprisMediaSession updateMediaNotification function
which then calls updateCurrentPlayer which used to call
SystemVolumeProvider's startTrackingVolumeKeys (now renamed to
startListeningForSinks) which is quite bad because on each call:
 * It would add yet another copy of this to the listeners list
 * It would send a network packet to get the sinks from the other device
Now we only call startListeningForSinks when the tracked player changes.