Add fastlane
This is read by F-Droid, so it should fix our app not having an icon or
having outdated screenshots there.

Also, this can be used to update the Play Store by doing:
  fastlane supply --version-code <version code> --json-key <path to key>

In a future, it would be nice to update the StaticMessages.sh script so
it converts the app description here to .pot and back so we can have KDE
localization teams translate it and updates are automatically picked up.