Use allowed CI image
## Summary

A recent (couple of weeks ago?) change to the Invent policies requires that we use a docker image from an allow list.

The cleanest fix is to just use the Ubuntu base image and install the JDK, which is approximately what we had before.

We considered the Sysadmin Android image (https://invent.kde.org/sysadmin/ci-images/-/blob/master/android-qt66/Dockerfile?ref_type=heads), but we use a newer version of the SDK tools and a newer version of the JDK, at which point we may as well just install those since those are the only dependencies. We could consider building another custom docker image in the future if we see value to doing so.

## Test Plan

### Before:

CI fails with:

```
ERROR: The "eclipse-temurin:17-jdk-focal" image is not present on list of allowed images:
- invent-registry.kde.org/sysadmin/ci-images/*:*
- ubuntu:*
- debian:*
- fedora:*
- centos:*
- opensuse/*:*
- python:*
- ruby:*
- fsfe/reuse:*
- quay.io/podman/stable:*
```

### After:

Build passes and drops .apk