Add way to send an error when trying to use a plugin that lacks permissions
Plugins can opt-in to be loaded even when missing permissions, so they can
react if the user tries to remotely use the plugin for example to show an
error.

This is used in the SFTP plugin to display a message when trying to use it
without having granted filesystem permissions.

I also renamed the name of the plugin shown when missing permissions, from
"filesystem expose" to something hopefully more understandable.